Unfinished 06 to GLVL Converter
Any object that needs Maths done to it (volumes for example) don't convert right, being 100x larger than they should be.

Blame GLVL for having a lot of hardcoding.
