+++ Tools/GLVL Converter/GLVL Converter/App.config
@@ -20,6 +20,9 @@
                 <setting name="lastSavedOutput" serializeAs="String">
                         <value />
                 </setting>
+                <setting name="s06toGLVLCheck" serializeAs="String">
+                        <value>False</value>
+                </setting>
         </GLvl_Converter.Properties.Settings>
         <GLVL_Converter.Properties.Settings>
                 <setting name="glvlTemplates" serializeAs="String">
