+++ Tools/GLVL Converter/GLVL Converter/Properties/Settings.settings
@@ -11,5 +11,8 @@
     <Setting Name="lastSavedOutput" Type="System.String" Scope="User">
       <Value Profile="(Default)" />
     </Setting>
+    <Setting Name="s06toGLVLCheck" Type="System.Boolean" Scope="User">
+      <Value Profile="(Default)">False</Value>
+    </Setting>
   </Settings>
 </SettingsFile>
