+++ Tools/GLVL Converter/GLVL Converter/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="DumpExtensions" version="2.0.0" targetFramework="net472" />
+</packages>
